date adjustment methods added
isBankHoliday()
Checks date passed with hardcoded bankholidays

checkBonusDate($date)
Checks if passed date falls on W/E or bank holiday if so passes the next tuesday.  If not the original date returned

checkPaymentDate($date)
Checks if passed date falls on W/E or bank holiday if so passes the previous friday.  If not the original date returned